Feat : feature/store_CRUD 20230913 1755

+++ Front-End/pro/app/src/main/java/com/example/project/AuctionPage.kt
@@ -98,9 +98,9 @@ fun AuctionPage(navController: NavHostController) {
                     filter2Selected = "소분류"
                     viewModel.applyFilter(
                         AuctionFilterDTO(
-                            filter1Selected,
-                            filter2Selected,
-                            filter3Selected,
+                            filter1Selected.toInt(),
+                            filter2Selected.toInt(),
+                            filter3Selected.toInt(),
                             searchText,
                             currentPage
                         )
@@ -119,9 +119,9 @@ fun AuctionPage(navController: NavHostController) {
                     filter2Selected = it
                     viewModel.applyFilter(
                         AuctionFilterDTO(
-                            filter1Selected,
-                            filter2Selected,
-                            filter3Selected,
+                            filter1Selected.toInt(),
+                            filter2Selected.toInt(),
+                            filter3Selected.toInt(),
                             searchText,
                             currentPage
                         )
@@ -135,9 +135,9 @@ fun AuctionPage(navController: NavHostController) {
                     filter3Selected = it
                     viewModel.applyFilter(
                         AuctionFilterDTO(
-                            filter1Selected,
-                            filter2Selected,
-                            filter3Selected,
+                            filter1Selected.toInt(),
+                            filter2Selected.toInt(),
+                            filter3Selected.toInt(),
                             searchText,
                             currentPage
                         )
